Fixed bug where banned routes were displaying incorrectly.
When changing the discipline of a route from banned to another discipline, the route would still show as banned. This is because a banned route was was tagged in the database with a banned field, but should have been marked by the discipline. This has been changed. Banned field in database no longer required.